refactor: extract BuildingValidationTools.cs
